Refactor Alias as TS
